Adding Callable class.
This is going to be the base class for built-in and user-defined
functions, so that the upcoming implicit casting functionality can be
used for calling both kind of functions without having to be repeated.

The AddIntegers, AddFloats, and AddDoubles were also changed to derive
from the Callable class. We might need to add a class BuiltInFunction
to be the base of these classes and all built-in functions, but for now,
I can't think of a use for it.

The Function class should also derive from the Callable class, and
probably be renamed to UserDefinedFunction. If we do the rename, it
"might" be better to rename Callable to Function.