Change GenerateCode() to save the generated value instead of returning it.
That is, GenerateCode() doesn't return the generated LLVM value.
Instead, it saves it internally and when needed, it can be retrieved via
the GetGeneratedLlvmValue() member function. This change is required
in case we want to allow other classes to use the generated LLVM value,
yet they shouldn't be calling the GenerateCode() member function
themselves.

For example, a CallFunction instruction might ask the callee to create
the necessary LLVM instruction (because CallFunction doesn't know what
the callee is, so it can't create the instruction itself). The callee
will need to access the generated value of the arguments passed to
CallFunction, yet it shouldn't be responsible for generated them;
CallFunction should be. Thus CallFunction would generate the code for
the arguments and the callee can use the generated value.