Support calling built-in functions in CallFunction.
This is so that we can implement built-in functions for arithmetic
operators, and since CallFunction was recently changed to support
implicit casting, this would allow us to call arithmetic operators on
values of different types.

The implementation through the addition of a Function base class which
serves as the base for both built-in functions and user defined
functions. Then CallFunction was changed such that it deals with
Function class rather than UserDefinedFunction.

In this commit, I also refactored a new class ExternalFunction out of
the UserDefinedFunction as the latter was serving both of them, which
is not ideal.

Finally, the member function Expression::GetValueType() was changed to
GetValueTypeSpec() and it is now returning a ValueTypeSpec instead of
a ValueType. This was necessary because otherwise we would have needed
a Module object to be able to get the ValueType out of the type specs
passed to Function (built-in, user defined, and external) objects.