Support expressions with multiple binary operators.
This was discovered while implementing a program from the book "The C
Programming Language", and failing to correctly parse the expression:

celsius = 5 * (fahr-32) / 9;

After inspection, it turned out that the "/ 9" part was being ignored
because CodeGenerator::GenerateBinaryOperator() only took the
first operator.

This commit also includes the the implementation of the first program
in section 1.2 in the book mentioned above. It is not currently being
automatically executed in the end-to-end tests because I decided to
change the structure of end-to-end tests a little bit to avoid having
a large number of programs in the same directory.