Restructing end-to-end test files.
The current directories are:
Arabic
General
TheCProgrammingLanguage

Also, to avoid having to do 'make install' before we can do 'make test',
I changed the E2E's CMake file such that it points to tests files from
the source directory instead of the binary directory. I also removed the
command to copy the files to the binary directory because it is not
needed any more. Finally, during the copy to the Examples directory,
I am copying the directory as they are to keep the hierarchy, rather
then dumping tens (and hundreds soon) to the Examples directory and
make it hard to find what we need.