Implementing a general Get() methods in value types' classes.
This is so that we can re-use types instead of defining a new type
each time it is requested, even if it has been used before. This:

1. Reduces the memory used for data types.
2. Allow easy and fast comparison of types by simply comparing their
   memory locations.

To make sure that users always use the Get() methods we use, the
constructors of value types' classes have been changed to privat.