Implement automatic type deduction.
This includes changing the DefineVariable instruction to support
accepting an expression that decides the type of the new variable. Also,
CodeGenerator.cpp was changed so that it can accepts expressions of
the form:

def <var> := <value>;

An end-to-end test type_deduction.alusus was added to test type
deduction with basic data types.