Support building Alusus on macOS
I made the necessary changes to be able to build on macOS. I used macOS
Mojave 10.14.6 (18G3020) and Clang with Apple LLVM version 10.0.1
(clang-1001.0.46.4).

I had to make multiple changes to make it work. First, the build script
was assuming the dynamic library to always have the `.so` extension,
which is not the case for macOS; macOS uses `.dylib` extension.

The second change I had to do was removing the `--enable-new-dtags` from
the linker arguments as the Clang on my machine complained about this
argument. Furthermore, I had to include `unordered_map` in
`LexerModule.h` as I experienced compilation errors.

The final change I had to make was in relation to `rpath` as it is
handled differently on macOS than on Linux. This is explained well in
[this
article](https://blogs.oracle.com/dipol/dynamic-libraries,-rpath,-and-mac-os)
and [this
article](https://gitlab.kitware.com/cmake/community/-/wikis/doc/cmake/RPATH-handling)
but briefly to have macOS looks in the rpath directories, the dynamic
library has to be referenced like @rpath/libSomething.dylib. When macOS
sees @rpath it will search for libSomething in the following directories
in order:

- DYLD_LIBRARY_PATH
- RPATH
- builtin directories
- DYLD_FALLBACK_LIBRARY_PATH

For more details, you can refer to the articles I linked above.

Here are the tests I did to confirm the rpath configuration in the
executable and dynamic libraries are correct:

```
otool -L ./Bin/alusus.dbg-0.6.1
./Bin/alusus.dbg-0.6.1:
	@rpath/libalusus_storage.0.6.dylib (compatibility version 0.6.0, current version 0.6.1)
	/usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 1252.250.1)
	/usr/lib/libc++.1.dylib (compatibility version 1.0.0, current version 400.9.4)
```

As you can see, the `libalusus_storage` is searched under the
rpath. Furthermore, executing `otool -l` and grepping for rpaths, I
could see the right paths added:

```
otool -l ./Bin/alusus.dbg-0.6.1 | grep LC_RPATH -A2
          cmd LC_RPATH
      cmdsize 32
         path @executable_path (offset 12)
--
          cmd LC_RPATH
      cmdsize 32
         path @executable_path (offset 12)
--
          cmd LC_RPATH
      cmdsize 40
         path @executable_path/../Lib/ (offset 12)
```

Notice that macOS doesn't support the `$ORIGIN` so instead I used the
`@executable_path` token which refers to where the executable path
lives, which in this case refers to the `Bin` directory. Also notice
that @executable_path is added twice, which I believe is because the
root `CMakeLists.txt` file also add the @executable_path (or $ORIGIN in
case of Linux) to the rpath flag of the linker, in addition to the one
being added by `Core/CMakeLists.txt`. I think this might be the case
whether Linux though I haven't tested it.

As for `libalusus_spp` and `libalusus_storage` executing `otool -D` on
them confirm they have the `@rpath` prefix:

```
otool -D ./Lib/libalusus_spp.dbg.0.6.1.dylib
./Lib/libalusus_spp.dbg.0.6.1.dylib:
@rpath/libalusus_spp.dbg.0.6.dylib

otool -D ./Lib/libalusus_storage.0.6.dylib
./Lib/libalusus_storage.0.6.dylib:
@rpath/libalusus_storage.0.6.dylib
```

Finally, I confirmed that I was able to execute `alusus.dbg` without
having to be under the `Lib folder:

```
./Bin/alusus.dbg
Alusus Language
Version 0.6.1-GIT902af07 (2020-04-19)
Copyright (C) 2020 Sarmad Khalid Abdullah
```