Enabled using blocks as function argument groups.
This allows using the block syntax {} to provide a list of arguments
without having to manually provide the count of the arguments. The
code generator automatically converts the block into count followed
by all elements in the block. This is useful in variadic functions where
the count of arguments is variable.