Implemented temp_ref.
Temp refs are used to allow automatic generation of temp vars if we have
a value and we are trying to cast it to a reference of the same type. If
the reference type is temp then a temp var will be created. If the
reference type is not temp then an error will be reported.