Fix the padding for AST text based types.
Text based AST types in Alusus used to pad the type with a
fixed size array to account for extra members that are internal
to the type and used on the C++ size, but of no use on Alusus
side. This causes breakage when running the code on other
platforms or when building with a different compiler.
This fix generates the padding dynamically depending on
actual type size info computed at preprocess time.