Removed function signature pre-processing.
Removed the automatic preprocessing of member function signature in
favour of using the `handler` command. Defining member functions using
the `func` command now requires using the @member attribute and
manually defining the `this` argument. This is to simplify the `func`
command and remove ambiguity around  the command behaving
differently depending on the context.