Fix bug of closures inside templates.
Fix a bug where closure argument types could not be found
when the type is a template argument.
Fixed by introducing Variable AST type which will later be
used for other source code clean ups. Closures now use this
new AST type to point directly to the type instead of cloning
the reference which does not work when the reference is
attached to a different scope.