Limit auto casting of smart refs to inheritance.
Update smart refs to not auto cast to the type of the
first member variable unless if that member var is
defined as an injection.
Also auto casting is no longer added if the target type
is the same as the original type (ex: a class whose first
member is a SrdRef of the same type as that class).