Fixed a bug with logical operators.
Fixed a bug with logical operators causing a seg fault
when the second operand use temp vars. The seg
fault is caused by the temp var being terminated in the
main context rather than the context of the second
operand, resulting in potential termination of an
un-initialized var in cases where execution of the
second oeprand is skipped.